HW-1 어댑터에 있던 데이터호출을 Fragment 안으로 옮겼습니다
