+++ preludezdev/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/ui/MarketFragment.kt
@@ -1,15 +1,19 @@
 package com.example.architecturestudy
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.architecturestudy.data.CoinTickerResponse
+import com.example.architecturestudy.network.NetworkHelper
 import com.example.architecturestudy.ui.MainActivity
 import com.example.architecturestudy.ui.RecyclerViewAdapter
 import kotlinx.android.synthetic.main.fragment_market.*
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
 class MarketFragment : Fragment() {
 
@@ -46,10 +50,45 @@ class MarketFragment : Fragment() {
         var type = arguments?.get("type")
 
         when (type) {
-            "KRW" -> rvAdapter.loadData(MainActivity.KRW_MARKETS)
-            "BTC" -> rvAdapter.loadData(MainActivity.BTC_MARKETS)
-            "ETH" -> rvAdapter.loadData(MainActivity.ETH_MARKETS)
-            "USDT" -> rvAdapter.loadData(MainActivity.USDT_MARKETS)
+            "KRW" -> loadData(MainActivity.KRW_MARKETS)
+            "BTC" -> loadData(MainActivity.BTC_MARKETS)
+            "ETH" -> loadData(MainActivity.ETH_MARKETS)
+            "USDT" -> loadData(MainActivity.USDT_MARKETS)
         }
+    }
+
+    fun loadData(markets: HashSet<String>) {
+        rvAdapter.clearData()
+
+        var marketList = StringBuffer()
+
+        for (market in markets) {
+            marketList.append(market + ",")
+        }
+        marketList.deleteCharAt(marketList.length - 1)
+
+        //Ticker 정보들 가져오기
+        NetworkHelper
+            .coinApiService
+            .getCurrTicker(marketList.toString())
+            .enqueue(object : Callback<List<CoinTickerResponse>> {
+                override fun onFailure(call: Call<List<CoinTickerResponse>>, t: Throwable) {
+                    t.printStackTrace()
+                }
+
+                override fun onResponse(
+                    call: Call<List<CoinTickerResponse>>,
+                    response: Response<List<CoinTickerResponse>>
+                ) {
+                    if (response.isSuccessful) {
+                        var list = response.body()
+                        for (ticker in list!!) {
+                            rvAdapter.addData(ticker)
+                        }
+
+                        rvAdapter.notifyDataSetChanged()
+                    }
+                }
+            })
     }
 }
